api/requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-cors-headers==2.2.0
django-environ==0.4.4
django-extensions==1.9.8
django-filter==1.1.0
django-simple-history==2.0
djangorestframework==3.7.7
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
flake8==3.5.0
idna==2.6
mccabe==0.6.1
oauthlib==2.0.6
pep8==1.7.1
psycopg2==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
typing==3.6.2
urllib3==1.22
Werkzeug==0.14.1
django_countries==5.3.1